Use correct pg_config for Debian package build
Before, it would use whatever pg_config was in the path, which was not
necessarily the one that belonged to the PostgreSQL major version for
which the package was meant to be built.  That could result in broken
or invalid packages.

Fix that by allowing overriding the pg_config location via a make
variable, and make the yada setup use this.