pgmemcache: stop using postgresql memory allocation in libmemcached
Libmemcached (and different versions of it) may or may not use the memory
allocation functions given to it with memcached_set_memory_allocators which
causes segmentation faults when calls to postgresql malloc/free are mixed
with calls to libc malloc/free.

It turns out that libmemcached calls from pgmemcache doesn't really need to
use postgresql's memory allocation functions as we never pass the data as-is
back to pg anyway, and we can just manually free the couple of temporary
allocations that are created.