Remove old servers from context when updating default_servers
libmemcached does not have a function for removing servers from the current
context so we must recreate the context from scratch when the list of
default servers changes.

The new functionality also makes it possible to implement a
memcache_server_delete() function at some point as requested in [#4] but
it's not implemented in this commit.