delta operations: fix negative offsets in increment and decrement
Memcached defines increment and decrement functions with uint64_t offsets.
PostgreSQL only has signed types, so we had some code that was supposed to
throw an error if a negative offset was passed in.  But the variable it was
looking at was unsigned, so the comparison was never true.

Instead of just fixing it to use a signed operation, let's invert the
operation if negative offset is passed in, effectively turning increment(-1)
into decrement(1).