behaviors: compare full behavior name, not just prefixes [#28]
get_memcached_behavior_flag et al were using strncmp to compare flags for an
unknown reason.  It caused flags with common prefixes to be compared
incorrectly, for example HASH_WITH_PREFIX_KEY was interpreted as HASH.

Revert to strcmp() instead which simplifies the code and fixes the issue.

Thanks olegdunkan for reporting the issue and suggesting a patch.