memcache_get_multi: refactor the code and fix warnings
Use DatumGetTextP(elem) to read text values from array instead of accessing
the Datums directly.  This silences a compiler warning and is the right
thing to do although it's unlikely that this makes a difference in practice
as keys are limited in size anyway (max 250 characters) and thus aren't
likely to be toasted, and DatumGetTextP() is basically just a cast for
non-toasted values.

Reduce #ifdef USE_OMCACHE/USE_LIBMEMCACHED hackery a bit and just copy the
loop separately for both implementations.  Use const unsigned char ** for
keys in OMcache and const char ** for keys in libmemcached.